--- src/pages/Home.tsx
@@ -1,5 +1,0 @@
-export const Home = () => {
-  return (
-    <main data-testid='home'>Home</main>
-  )
-};
