+++ src/pages/Home/Home.tsx
@@ -1,0 +1,9 @@
+import { Hero } from "./components/Hero/Hero";
+
+export const Home = () => {
+  return (
+    <main data-testid='home'>
+      <Hero />
+    </main>
+  )
+};
