+++ src/pages/index.js
@@ -1,2 +1,2 @@
-export { Home } from './Home';
+export { Home } from './Home/Home';
 export { PageNotFound } from './PageNotFound/PageNotFound';
